requirements.txt
langgraph
langchain
google-generativeai
python-dotenv
pandas
yfinance
gradio